50. Pow(x, n) using different approaches

+++ Recursion/PowXN.java
@@ -41,7 +41,7 @@ public class PowXN {
     // Space Complexity: O(1)
     public static double myPow(double x, int n) {
         double result = 1.0;
-        long long nn = n; // Handle negative powers by converting n to long
+        long nn = n; // Handle negative powers by converting n to long
         if (nn < 0) nn = -1 * nn;
         while (nn > 0) {
             if (nn % 2 == 1) {
